refactor(analysis): split job claim modules and include hdbscan changes
